Implement formatter (#33)
