+++ cmd/regex_completion_test.go
